Player Collision Fix
Update-
PlayerContorl: if player not holding A or D the velocity will be set to zero.

JumpController: made velocity of z axis to 0 upon jump.
CollisionDetector: in on collision enter made force applied to player equal through collision contact normal.

Tags Created: Wall
Created to make the player detect on collision exit so the player's velocity on z gets set to 0 on exit if 'wall'.